@@ -9,16 +9,17 @@ using System.Linq.Expressions;
 using System.Reflection;
 using System.Windows.Forms;
 using System.Collections;
+using System.Diagnostics;
 
 namespace CFG
 {
-	public enum CFG_SpriteType
-	{
-		Normal = 0,
-		Custom = 1,
+    public enum CFG_SpriteType
+    {
+        Normal = 0,
+        Custom = 1,
         [DisplayName("Generator/Shooter")]
-		GeneratorShooter = 3,
-	}
+        GeneratorShooter = 3,
+    }
 
     public enum FileType
     {
@@ -40,16 +41,16 @@ namespace CFG
         Size16x16 = 16,
     }
 
-	public partial class CFG_Editor : Form
-	{
+    public partial class CFG_Editor : Form
+    {
         #region Properteis
 
         private string filename = "";
-		public string Filename 
-		{
-			get { return filename; }
-			set
-			{
+        public string Filename
+        {
+            get { return filename; }
+            set
+            {
                 if (filename != value)
                 {
                     filename = value;
@@ -58,8 +59,8 @@ namespace CFG
                     else
                         this.Text = "CFG Editor";
                 }
-			}
-		}
+            }
+        }
 
         private bool unsaved = false;
         public bool Unsaved
@@ -77,11 +78,11 @@ namespace CFG
 
         public readonly CfgFile Data;
 
-        		
-		public Image[] objClip = new Image[0x10];
-		public Image[] sprClip = new Image[0x40];
-		public Bitmap[] sprPal = new Bitmap[6 + 2 * 8];
-        
+
+        public Image[] objClip = new Image[0x10];
+        public Image[] sprClip = new Image[0x40];
+        public Bitmap[] sprPal = new Bitmap[6 + 2 * 8];
+
         /// <summary>
         /// The sprite palette to be used. Not the one from the CFG file but the one displayed when pal E or F are selected.
         /// </summary>
@@ -92,14 +93,26 @@ namespace CFG
         readonly BindingList<ComboBoxItem> types_list = new BindingList<ComboBoxItem>();
         readonly BindingList<ComboBoxItem> sprites_list = new BindingList<ComboBoxItem>();
         readonly Map16Resources resources = new Map16.Map16Resources();
-        
+
         FileType FileType;
         byte[] RomData = null;
         readonly List<ControlEnabler> control_enablers = new List<ControlEnabler>();
 
-		public CFG_Editor(string[] args)
+        public void DisplaySourceCurrentChanged(object o, EventArgs e)
         {
-            if(args.Length > 0)
+            spriteEditor1.Sprite = (DisplaySprite)displaySpriteBindingSource.Current;
+        }
+
+        public void CheckedExtraByteChanged(object _, EventArgs __)
+        {
+            Data.DispType = chbExtraByte.Checked ? Json.DisplayType.ExtraByte : Json.DisplayType.XY;
+            foreach (var entry in Data.DisplayEntries)
+                entry.disp_type = Data.DispType;
+        }
+
+        public CFG_Editor(string[] args)
+        {
+            if (args.Length > 0)
             {
                 int flag = Array.IndexOf(args, "-c");
                 if (flag >= 0)
@@ -112,6 +125,11 @@ namespace CFG
 
 
             InitializeComponent();
+            dgvGFXInfo.Columns[0].ReadOnly = false;
+            dgvGFXInfo.Columns[1].ReadOnly = false;
+            dgvGFXInfo.Columns[2].ReadOnly = false;
+            dgvGFXInfo.Columns[3].ReadOnly = false;
+            dgvGFXInfo.Columns[4].ReadOnly = false;
 
             #region map16 resources
 
@@ -125,7 +143,7 @@ namespace CFG
 
             byte[] map16data = new byte[0x2000];
             Array.Copy(CFG.Properties.Resources.m16Page1_3, map16data, 0x1800);
-            
+
             #endregion
 
 
@@ -170,24 +188,35 @@ namespace CFG
                 Unsaved = true;
                 if (e.PropertyName == nameof(Data.Type))
                     control_enablers.ForEach(c => c.Evaluate());
+                if (e.PropertyName == nameof(Data.DispType))
+                {
+                    chbExtraByte.CheckedChanged -= CheckedExtraByteChanged;
+                    chbExtraByte.Checked = Data.DispType == Json.DisplayType.ExtraByte;
+                    dgvDisplay.Columns[1].HeaderText = Data.DispType == Json.DisplayType.ExtraByte ? "Index" : "X";
+                    dgvDisplay.Columns[2].HeaderText = Data.DispType == Json.DisplayType.ExtraByte ? "Value" : "Y";
+                    label8.Text = Data.DispType == Json.DisplayType.ExtraByte ? "I:" : "X:";
+                    label9.Text = Data.DispType == Json.DisplayType.ExtraByte ? "V:" : "Y:";
+
+                    chbExtraByte.CheckedChanged += CheckedExtraByteChanged;
+                }
             };
 
             #region Default Tab
 
 
             cmb_1656_0F.BitsBind(Data, c => c.SelectedIndex, f => f.Addr1656, 4, 0).DataSourceUpdateMode = DataSourceUpdateMode.OnPropertyChanged;
-			cmb_1662_3F.BitsBind(Data, c => c.SelectedIndex, f => f.Addr1662, 6, 0).DataSourceUpdateMode = DataSourceUpdateMode.OnPropertyChanged;
-			cmb_166E_0E.BitsBind(Data, c => c.SelectedIndex, f => f.Addr166E, 3, 1).DataSourceUpdateMode = DataSourceUpdateMode.OnPropertyChanged;
+            cmb_1662_3F.BitsBind(Data, c => c.SelectedIndex, f => f.Addr1662, 6, 0).DataSourceUpdateMode = DataSourceUpdateMode.OnPropertyChanged;
+            cmb_166E_0E.BitsBind(Data, c => c.SelectedIndex, f => f.Addr166E, 3, 1).DataSourceUpdateMode = DataSourceUpdateMode.OnPropertyChanged;
 
-			SetupBinding(Data, f => f.Addr1656, "1656", 0);
-			SetupBinding(Data, f => f.Addr1662, "1662", 0);
-			SetupBinding(Data, f => f.Addr166E, "166E", 0);
-			SetupBinding(Data, f => f.Addr167A, "167A", 0);
-			SetupBinding(Data, f => f.Addr1686, "1686", 0);
-			SetupBinding(Data, f => f.Addr190F, "190F", 0);
-            
-			txt_0001.Bind(Data, c => c.Text, f => f.ExProp1, f => f.ToString("X2"), c => StringToByte(c));
-			txt_0002.Bind(Data, c => c.Text, f => f.ExProp2, f => f.ToString("X2"), c => StringToByte(c));
+            SetupBinding(Data, f => f.Addr1656, "1656", 0);
+            SetupBinding(Data, f => f.Addr1662, "1662", 0);
+            SetupBinding(Data, f => f.Addr166E, "166E", 0);
+            SetupBinding(Data, f => f.Addr167A, "167A", 0);
+            SetupBinding(Data, f => f.Addr1686, "1686", 0);
+            SetupBinding(Data, f => f.Addr190F, "190F", 0);
+
+            txt_0001.Bind(Data, c => c.Text, f => f.ExProp1, f => f.ToString("X2"), c => StringToByte(c));
+            txt_0002.Bind(Data, c => c.Text, f => f.ExProp2, f => f.ToString("X2"), c => StringToByte(c));
             nudNormal.Bind(Data, c => c.Value, f => f.ByteCount, null, c => (byte)c);
             nudExtra.Bind(Data, c => c.Value, f => f.ExByteCount, null, c => (byte)c);
             txtASMFile.Bind(Data, c => c.Text, f => f.AsmFile);
@@ -203,33 +232,33 @@ namespace CFG
 
             cmbType.DataSource = types_list;
 
-			//fetch clipping images from resources and add them to the comboboxes.
-			for (int i = 0; i < objClip.Length; i++)
-			{
-				using (Stream str = System.Reflection.Assembly.GetExecutingAssembly().GetManifestResourceStream("CFG.Resources.ObjClipping." + i.ToString("X2") + ".png"))
-					objClip[i] = ResizeImg(pcbObjClipping.Size, Image.FromStream(str));
-				cmb_1656_0F.Items.Add(i.ToString("X2"));
-			}
-			for (int i = 0; i < sprClip.Length; i++)
-			{
-				using (Stream str = System.Reflection.Assembly.GetExecutingAssembly().GetManifestResourceStream("CFG.Resources.SprClipping." + i.ToString("X2") + ".png"))
-					sprClip[i] = ResizeImg(pcbSprClipping.Size, Image.FromStream(str));
-				cmb_1662_3F.Items.Add(i.ToString("X2"));
-			}
-			
-			//draw bitmap of palette file for sprites.
-			for (int i = 0; i < sprPal.Length; i++)
-			{
-				Bitmap b = new Bitmap(8 * 16, 16);
-				using(Graphics g = Graphics.FromImage(b))
-					for(int pal = 0; pal < 8; pal++)
-					{
+            //fetch clipping images from resources and add them to the comboboxes.
+            for (int i = 0; i < objClip.Length; i++)
+            {
+                using (Stream str = System.Reflection.Assembly.GetExecutingAssembly().GetManifestResourceStream("CFG.Resources.ObjClipping." + i.ToString("X2") + ".png"))
+                    objClip[i] = ResizeImg(pcbObjClipping.Size, Image.FromStream(str));
+                cmb_1656_0F.Items.Add(i.ToString("X2"));
+            }
+            for (int i = 0; i < sprClip.Length; i++)
+            {
+                using (Stream str = System.Reflection.Assembly.GetExecutingAssembly().GetManifestResourceStream("CFG.Resources.SprClipping." + i.ToString("X2") + ".png"))
+                    sprClip[i] = ResizeImg(pcbSprClipping.Size, Image.FromStream(str));
+                cmb_1662_3F.Items.Add(i.ToString("X2"));
+            }
+
+            //draw bitmap of palette file for sprites.
+            for (int i = 0; i < sprPal.Length; i++)
+            {
+                Bitmap b = new Bitmap(8 * 16, 16);
+                using (Graphics g = Graphics.FromImage(b))
+                    for (int pal = 0; pal < 8; pal++)
+                    {
                         Color c = resources.Palette[i][pal];
-						Rectangle rec = new Rectangle(16 * pal, 0, 16, 16);
-						g.FillRectangle(new SolidBrush(c), rec);
-					}
-				sprPal[i] = b;
-			}
+                        Rectangle rec = new Rectangle(16 * pal, 0, 16, 16);
+                        g.FillRectangle(new SolidBrush(c), rec);
+                    }
+                sprPal[i] = b;
+            }
 
             #endregion
 
@@ -260,29 +289,58 @@ namespace CFG
             dsSP3.Tag = 0x2000;
             dsSP4.Tag = 0x3000;
             cmbTilesets.SelectedIndex = 0;
-            
+
+            chbExtraByte.CheckedChanged += CheckedExtraByteChanged;
             //Grid size combobox data + events
             foreach (GridSize gridSize in Enum.GetValues(typeof(GridSize)).Cast<GridSize>().Reverse())
                 cmbGrid.Items.Add(new ComboBoxItem(gridSize.GetName(), (int)gridSize));
             cmbGrid.SelectedIndex = 0;
             cmbGrid.SelectedIndexChanged += (_, __) => spriteEditor1.GridSize = ((ComboBoxItem)cmbGrid.SelectedItem).Value;
-            
+
             //------------------------------------------------------------------------------------------------------
 
             //create binding source for later databinding.
             displaySpriteBindingSource.DataSource = Data;
             displaySpriteBindingSource.DataMember = nameof(CfgFile.DisplayEntries);
-            
+
+            gfxInfoBindingSource.DataSource = Data;
+            gfxInfoBindingSource.DataMember = nameof(CfgFile.GFXInfos);
+
+            dgvDisplay.AutoGenerateColumns = false;
+            dgvDisplay.Columns.Clear();
+            dgvDisplay.Columns.AddRange(new DataGridViewColumn[] {
+                new DataGridViewCheckBoxColumn
+                {
+                    HeaderText = "ExtraBit",
+                    Name = "ExtraBit",
+                    DataPropertyName = "ExtraBit"
+                },
+                new DataGridViewTextBoxColumn
+                {
+                    HeaderText = "X",
+                    Name = "X_or_index",
+                    DataPropertyName = "X_or_index"
+                },
+                new DataGridViewTextBoxColumn
+                {
+                    HeaderText = "Y",
+                    Name = "Y_or_value",
+                    DataPropertyName = "Y_or_value"
+                }
+            });
+            dgvGFXInfo.DataError += DgvGFXInfo_DataError;
+
             //create databindings between display list and controls.
             BindToSourceDisplay(rtbDesc, displaySpriteBindingSource, ctrl => ctrl.Text, ds => ds.Description);
-            BindToSourceDisplay(nudX, displaySpriteBindingSource, ctrl => ctrl.Value, ds => ds.X);
-            BindToSourceDisplay(nudY, displaySpriteBindingSource, ctrl => ctrl.Value, ds => ds.Y);
+            BindToSourceDisplay(nudX, displaySpriteBindingSource, ctrl => ctrl.Value, ds => ds.X_or_index);
+            BindToSourceDisplay(nudY, displaySpriteBindingSource, ctrl => ctrl.Value, ds => ds.Y_or_value);
             BindToSourceDisplay(chbExtraBit, displaySpriteBindingSource, ctrl => ctrl.Checked, ds => ds.ExtraBit);
             BindToSourceDisplay(chbUseText, displaySpriteBindingSource, ctrl => ctrl.Checked, ds => ds.UseText).DataSourceUpdateMode = DataSourceUpdateMode.OnPropertyChanged;
-            displaySpriteBindingSource.CurrentChanged += (_, __) => spriteEditor1.Sprite = (DisplaySprite)displaySpriteBindingSource.Current;
-
             ConnectViewAndButtons(displaySpriteBindingSource, dgvDisplay, btnDisplayNew, btnDisplayClone, btnDisplayDelete);
+            displaySpriteBindingSource.CurrentChanged += DisplaySourceCurrentChanged;
 
+            ConnectViewAndButtons(gfxInfoBindingSource, dgvGFXInfo, btnGFXNew, btnGFXClone, btnGFXRemove);
+            dgvGFXInfo.CellFormatting += DgvGFXInfo_CellFormatting;
 
             map16Editor1.Initialize(map16data, resources);
             map16Editor1.SelectionChanged += (_, e) => pnlEdit.Enabled = e.Tile >= 0x300;
@@ -313,7 +371,7 @@ namespace CFG
                 cmbPalette.SelectedIndex = map16Editor1.SelectedObject.Palette;
                 trigger = 0;
             };
-            
+
             btnX.Click += (_, __) => map16Editor1.SelectedObject.FlipX();
             btnY.Click += (_, __) => map16Editor1.SelectedObject.FlipY();
 
@@ -393,8 +451,47 @@ namespace CFG
             testToolStripMenuItem.Visible = true;
 #endif
 
-		}
-        
+        }
+
+        private void DgvGFXInfo_DataError(object sender, DataGridViewDataErrorEventArgs e)
+        {
+            e.Cancel = true;
+            MessageBox.Show("Invalid value for GFX info, only base 10 numbers are valid\nThey'll get converted to hex automatically", "Error", MessageBoxButtons.OK);
+        }
+
+        private void DgvGFXInfo_CellFormatting(object sender, DataGridViewCellFormattingEventArgs e)
+        {
+            var dgv = sender as DataGridView;
+            if (e.ColumnIndex != 0 && e.Value != null)
+            {
+                string v = e.Value.ToString();
+                if (v.StartsWith("0x") || v.StartsWith("0X"))
+                {
+                    if (int.TryParse(v.Substring(2), System.Globalization.NumberStyles.HexNumber, System.Globalization.CultureInfo.CurrentCulture, out int value))
+                    {
+                        e.Value = $"0x{(value > 0x7F || value < 0 ? 0x7F : value):X02}";
+                    }
+                    else
+                    {
+                        e.Value = "0x7F";
+                    }
+                }
+                else
+                {
+                    if (int.TryParse(v, out int value))
+                    {
+                        e.Value = $"0x{(value > 0x7F || value < 0 ? 0x7F : value):X02}";
+                    }
+                    else
+                    {
+                        e.Value = "0x7F";
+                    }
+                }
+
+            }
+            e.FormattingApplied = true;
+        }
+
         /// <summary>
         /// Testbutton on the menu. Only available in debug mode.
         /// </summary>
@@ -409,25 +506,39 @@ namespace CFG
         void ConnectViewAndButtons(BindingSource source, DataGridView dgv, Button btnNew, Button btnClone, Button btnRemove)
         {
             btnRemove.Enabled = false;
+            btnClone.Enabled = false;
             dgv.AllowUserToDeleteRows = false;
-
+            source.AllowNew = true;
             //events to control the button/view behaviour for new/clone/delete
             btnNew.Click += (_, __) => source.AddNew();
             btnClone.Click += (_, __) => source.Add(((ICloneable)source.Current).Clone());
             btnRemove.Click += (_, __) => source.RemoveCurrent();
 
-            dgv.RowsAdded += (_, __) => 
-                dgv.AllowUserToDeleteRows = dgv.Rows.Count > 1;
-            dgv.RowsRemoved += (_, __) => 
-                dgv.AllowUserToDeleteRows = dgv.Rows.Count > 1;
+
+            dgv.RowsAdded += (_, __) =>
+            {
+                dgv.AllowUserToDeleteRows = dgv.Rows.Count >= 1;
+                btnClone.Enabled = dgv.Rows.Count >= 1;
+            };
+            dgv.RowsRemoved += (_, __) =>
+            {
+                dgv.AllowUserToDeleteRows = dgv.Rows.Count >= 1;
+                btnClone.Enabled = dgv.Rows.Count >= 1;
+            };
+
             dgv.AllowUserToDeleteRowsChanged += (_, __) =>
                 btnRemove.Enabled = dgv.AllowUserToDeleteRows;
+            
         }
 
         private Binding BindToSourceCollection<TCon, TProp>(TCon control, BindingSource source, Expression<Func<TCon, TProp>> controlMember, Expression<Func<CollectionSprite, TProp>> objectMember)
             where TCon : Control
             => BindToSource(control, source, controlMember, objectMember);
         private Binding BindToSourceDisplay<TCon, TProp>(TCon control, BindingSource source, Expression<Func<TCon, TProp>> controlMember, Expression<Func<DisplaySprite, TProp>> objectMember)
+            where TCon : Control
+            => BindToSource(control, source, controlMember, objectMember);
+
+        private Binding BindToSourceGFX<TCon, TProp>(TCon control, BindingSource source, Expression<Func<TCon, TProp>> controlMember, Expression<Func<GFXInfo, TProp>> objectMember)
             where TCon : Control
             => BindToSource(control, source, controlMember, objectMember);
 
@@ -543,20 +654,20 @@ namespace CFG
         /// <param name="expr">The expression that selects the Property to bind to from the CFGFile</param>
         /// <param name="addr">The string representation of the address, used for searching through the controlls</param>
         /// <param name="start_bit">For the CheckBoxes, this indicates where to start</param>
-		private void SetupBinding(CfgFile file, System.Linq.Expressions.Expression<Func<CfgFile, byte>> expr, string addr, int start_bit = 0)
-		{
-			TextBox txt = (TextBox)Controls.Find("txt_" + addr, true).FirstOrDefault();
-			if (txt == null)
-				return;
-			txt.Bind(file, c => c.Text, expr, f => f.ToString("X2"), c => StringToByte(c));
-			for(int i = start_bit; i < 8; i++)
-			{
-				CheckBox ch = (CheckBox)Controls.Find("chb_" + addr + "_" + (1 << i).ToString("X2"), true).FirstOrDefault();
-				if (ch == null)
-					continue;
-				ch.BitBind(file, c => c.Checked, expr, i).DataSourceUpdateMode = DataSourceUpdateMode.OnPropertyChanged;
-			}
-		}
+        private void SetupBinding(CfgFile file, System.Linq.Expressions.Expression<Func<CfgFile, byte>> expr, string addr, int start_bit = 0)
+        {
+            TextBox txt = (TextBox)Controls.Find("txt_" + addr, true).FirstOrDefault();
+            if (txt == null)
+                return;
+            txt.Bind(file, c => c.Text, expr, f => f.ToString("X2"), c => StringToByte(c));
+            for (int i = start_bit; i < 8; i++)
+            {
+                CheckBox ch = (CheckBox)Controls.Find("chb_" + addr + "_" + (1 << i).ToString("X2"), true).FirstOrDefault();
+                if (ch == null)
+                    continue;
+                ch.BitBind(file, c => c.Checked, expr, i).DataSourceUpdateMode = DataSourceUpdateMode.OnPropertyChanged;
+            }
+        }
 
         private void SetDataSelectorGfx(FileSelector ds, int id)
         {
@@ -592,13 +703,13 @@ namespace CFG
         }
 
         #endregion
-        
+
         public void LoadFile(string path)
         {
             if (!File.Exists(path))
                 throw new FileNotFoundException($"\"{path}\" not found.");
             string ext = Path.GetExtension(path);
-            switch(ext.ToLower())
+            switch (ext.ToLower())
             {
                 case ".cfg":
                     Data.FromLines(File.ReadAllText(path));
@@ -615,7 +726,7 @@ namespace CFG
                 default:
                     throw new FormatException($"Uknown file extension {ext}");
             }
-            
+
             if (FileType == FileType.CfgFile)
             {
                 cmbType.DataSource = types_list;
@@ -649,7 +760,7 @@ namespace CFG
             }
 
         }
-        
+
         public void SaveFile(string path)
         {
             //update binding in case the current control is a textbox,
@@ -680,27 +791,27 @@ namespace CFG
                 default:
                     throw new FormatException($"Uknown file extension {ext}");
             }
-            
+
         }
 
-		#region Events
-		private void Cmb_1656_0F_SelectedIndexChanged(object sender, EventArgs e)
-		{
-			lblObjBroke.Visible = ((ComboBox)sender).SelectedIndex >= 0x0F;
-			pcbObjClipping.Image = objClip[((ComboBox)sender).SelectedIndex];
-		}
-		private void Cmb_1662_3F_SelectedIndexChanged(object sender, EventArgs e)
-		{
-			lblSprBroke.Visible = ((ComboBox)sender).SelectedIndex >= 0x3C;
-			pcbSprClipping.Image = sprClip[((ComboBox)sender).SelectedIndex];
-		}
-		private void Cmb_166E_0E_SelectedIndexChanged(object sender, EventArgs e)
-		{
-			int index = ((ComboBox)sender).SelectedIndex;
-			if(index > 5)
-				index += SpritePalette * 2;
-			pcbPal.Image = sprPal[index];
-		}
+        #region Events
+        private void Cmb_1656_0F_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            lblObjBroke.Visible = ((ComboBox)sender).SelectedIndex >= 0x0F;
+            pcbObjClipping.Image = objClip[((ComboBox)sender).SelectedIndex];
+        }
+        private void Cmb_1662_3F_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            lblSprBroke.Visible = ((ComboBox)sender).SelectedIndex >= 0x3C;
+            pcbSprClipping.Image = sprClip[((ComboBox)sender).SelectedIndex];
+        }
+        private void Cmb_166E_0E_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            int index = ((ComboBox)sender).SelectedIndex;
+            if (index > 5)
+                index += SpritePalette * 2;
+            pcbPal.Image = sprPal[index];
+        }
 
         /// <summary>
         /// Only allow hex values (and backspace) to be inputed into the textboxes
@@ -742,31 +853,31 @@ namespace CFG
             saveAsToolStripMenuItem.Enabled = true;
             this.Filename = "";
             this.Unsaved = true;
-		}
+        }
 
-		private void LoadToolStripMenuItem_Click(object sender, EventArgs e)
-		{
+        private void LoadToolStripMenuItem_Click(object sender, EventArgs e)
+        {
             OpenFileDialog ofd = new OpenFileDialog
             {
                 Filter = "CFG or JSON file|*.json;*.cfg|ROM file|*.smc;*.sfc",
                 Title = "Load CFG file"
             };
             if (ofd.ShowDialog() == System.Windows.Forms.DialogResult.Cancel)
-				return;
-			try
-			{
-				LoadFile(ofd.FileName);
-			}
-			catch (Exception ex)
-			{
+                return;
+            try
+            {
+                LoadFile(ofd.FileName);
+            }
+            catch (Exception ex)
+            {
                 Data.Clear();
-				Filename = "";
-				MessageBox.Show("An error occured while trying to read the CFG data\n" + ex.Message, "Unexpected Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
-			}
-		}
+                Filename = "";
+                MessageBox.Show("An error occured while trying to read the CFG data\n" + ex.Message, "Unexpected Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+        }
 
-		private void SaveAsToolStripMenuItem_Click(object sender, EventArgs e)
-		{
+        private void SaveAsToolStripMenuItem_Click(object sender, EventArgs e)
+        {
             if (FileType == FileType.RomFile)
             {
                 MessageBox.Show("You can't save a ROM under a different filename. Sorry about that.", "Error",
@@ -781,21 +892,21 @@ namespace CFG
                 FileName = Path.GetFileName(Path.ChangeExtension(Filename, ".json"))
             };
             if (sfd.ShowDialog() == System.Windows.Forms.DialogResult.Cancel)
-				return;
+                return;
 
             try
             {
                 SaveFile(sfd.FileName);
                 Filename = sfd.FileName;
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 ShowException(ex);
             }
-		}
+        }
 
-		private void SaveToolStripMenuItem_Click(object sender, EventArgs e)
-		{
+        private void SaveToolStripMenuItem_Click(object sender, EventArgs e)
+        {
             if (Filename == "")
                 SaveAsToolStripMenuItem_Click(sender, e);
             else
@@ -809,7 +920,7 @@ namespace CFG
                     ShowException(ex);
                 }
             }
-		}
+        }
         #endregion
 
         private void Ds_FileLoaded(object sender, EventArgs e)
