@@ -8,12 +8,50 @@ using System.Text.RegularExpressions;
 using System.Drawing;
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
+using CFG.Json;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 
 namespace CFG.Map16
 {
-	[DebuggerDisplay("X={X},Y={Y},Ex={ExtraBit} Tiles: {Tiles.Count}")]
-	public class DisplaySprite : ICloneable, INotifyPropertyChanged
-	{
+    [DebuggerDisplay("X={X_or_index},Y={Y_or_value},Ex={ExtraBit} Tiles: {Tiles.Count}")]
+    public class DisplaySprite : ICloneable, INotifyPropertyChanged
+    {
+        [JsonIgnore]
+        private DisplayType _displayType;
+
+        [JsonIgnore]
+        public DisplayType disp_type
+        {
+            get
+            {
+                return _displayType;
+            }
+            set
+            {
+                _displayType = value;
+                if (_displayType == DisplayType.XY)
+                {
+                    if (_X == null && _Y == null)
+                    {
+                        _X = _Index.GetValueOrDefault();
+                        _Y = _Value.GetValueOrDefault();
+                        _Index = null;
+                        _Value = null;
+                    }
+                }
+                else
+                {
+                    if (_Index == null && _Value == null)
+                    {
+                        _Index = _X.GetValueOrDefault();
+                        _Value = _Y.GetValueOrDefault();
+                        _X = null;
+                        _Y = null;
+                    }
+                }
+            }
+        }
         /// <summary>
         /// The default sprite with only a description and one tile (a pink X)
         /// </summary>
@@ -25,9 +63,9 @@ namespace CFG.Map16
                 new Tile(0, 0, 0),
             }
         };
-        
+
         private string _Description = "";
-		public string Description
+        public string Description
         {
             get { return _Description; }
             set { SetPropertyValue(ref _Description, value); }
@@ -39,6 +77,7 @@ namespace CFG.Map16
             get { return _ExtraBit; }
             set { SetPropertyValue(ref _ExtraBit, value); }
         }
+
         private bool _CustomBit = true;
         [Newtonsoft.Json.JsonIgnore]
         public bool CustomBit
@@ -49,17 +88,61 @@ namespace CFG.Map16
 
         public BindingList<Tile> Tiles { get; set; }
 
-        private int _X = 0;
-		public int X
+        [Newtonsoft.Json.JsonIgnore]
+        public bool CheckDispType => disp_type == DisplayType.XY;
+
+        class DisplayTypeConverter : JsonConverter
         {
-            get { return _X; }
-            set { SetPropertyValue(ref _X, value); }
+            public override bool CanConvert(Type objectType)
+            {
+                return (objectType == typeof(int));
+            }
+
+            public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
+            {
+                var token = JToken.Load(reader);
+                if (token.Type == JTokenType.Integer)
+                {
+                    return (int)token;
+                }
+                return null;
+            }
+
+            public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
+            {
+                if (value != null)
+                    serializer.Serialize(writer, ((int?)value).Value);
+            }
         }
-        private int _Y = 0;
-        public int Y
+
+        [JsonProperty(PropertyName = "X", NullValueHandling = NullValueHandling.Ignore)]
+        [JsonConverter(typeof(DisplayTypeConverter))]
+        private int? _X = null;
+
+        [JsonProperty(PropertyName = "Y", NullValueHandling = NullValueHandling.Ignore)]
+        [JsonConverter(typeof(DisplayTypeConverter))]
+        private int? _Y = null;
+
+        [JsonProperty(PropertyName = "Value", NullValueHandling = NullValueHandling.Ignore)]
+        [JsonConverter(typeof(DisplayTypeConverter))]
+        private int? _Value = null;
+
+        [JsonProperty(PropertyName = "Index", NullValueHandling = NullValueHandling.Ignore)]
+        [JsonConverter(typeof(DisplayTypeConverter))]
+        private int? _Index = null;
+
+        [JsonIgnore]
+        public int X_or_index
         {
-            get { return _Y; }
-            set { SetPropertyValue(ref _Y, value); }
+            get { return (CheckDispType ? _X : _Index).GetValueOrDefault(); }
+            set { SetPropertyValue(ref CheckDispType ? ref _X : ref _Index, value); }
+        }
+
+        [JsonIgnore]
+        public int Y_or_value
+        {
+            get { return (CheckDispType ? _Y : _Value).GetValueOrDefault(); }
+            set { SetPropertyValue(ref CheckDispType ? ref _Y : ref _Value, value); }
         }
 
         private string _DisplayText = "";
@@ -86,7 +169,7 @@ namespace CFG.Map16
         //    sprite.Y = Convert.ToInt32(tilesLine.Substring(0, 1), 16);
         //    sprite.X = Convert.ToInt32(tilesLine.Substring(1, 1), 16);
         //    sprite.ExtraBit = tilesLine[2] == '3';
-            
+
         //    foreach (Tile t in Tile.GetTiles(tilesLine))
         //        sprite.Tiles.Add(t);
         //}
@@ -94,78 +177,78 @@ namespace CFG.Map16
         #region Object Overrides
 
         public override string ToString()
-		{
-			int ex = (CustomBit ? 2 : 0) + (ExtraBit ? 1 : 0);
-			return ex + ":" + Y.ToString("X") + X.ToString("X");
-		}
-		public override bool Equals(object obj)
-		{
-			if (object.ReferenceEquals(this, obj))
-				return true;
-			if (object.ReferenceEquals(this, null))
-				return false;
+        {
+            int ex = (CustomBit ? 2 : 0) + (ExtraBit ? 1 : 0);
+            return ex + ":" + Y_or_value.ToString("X") + X_or_index.ToString("X");
+        }
+        public override bool Equals(object obj)
+        {
+            if (object.ReferenceEquals(this, obj))
+                return true;
+            if (object.ReferenceEquals(this, null))
+                return false;
 
-			DisplaySprite sp = obj as DisplaySprite;
-			if (object.ReferenceEquals(sp, null))
-				return false;
+            DisplaySprite sp = obj as DisplaySprite;
+            if (object.ReferenceEquals(sp, null))
+                return false;
 
             return sp.CustomBit == CustomBit && sp.ExtraBit == ExtraBit
                 && sp.Description == Description && sp.DisplayText == DisplayText && sp.UseText == UseText
                 && sp.Tiles.SequenceEqual(Tiles);
-		}
-		public override int GetHashCode()
-		{
-			return (ExtraBit ? 0x100 : 0) + (CustomBit ? 0x200 : 0) + (X ^ Y);
-		}
+        }
+        public override int GetHashCode()
+        {
+            return (ExtraBit ? 0x100 : 0) + (CustomBit ? 0x200 : 0) + (X_or_index ^ Y_or_value);
+        }
 
-		#endregion
-		#region Operators
+        #endregion
+        #region Operators
 
-		public static bool operator ==(DisplaySprite sp1, DisplaySprite sp2)
-		{
-			if (object.ReferenceEquals(sp1, sp2))
-				return true;
-			if (object.ReferenceEquals(sp1, null))
-				return false;
-			if (object.ReferenceEquals(sp2, null))
-				return false;
+        public static bool operator ==(DisplaySprite sp1, DisplaySprite sp2)
+        {
+            if (object.ReferenceEquals(sp1, sp2))
+                return true;
+            if (object.ReferenceEquals(sp1, null))
+                return false;
+            if (object.ReferenceEquals(sp2, null))
+                return false;
 
-			return sp1.CustomBit == sp2.CustomBit && sp1.ExtraBit == sp2.ExtraBit;
-		}
-		public static bool operator !=(DisplaySprite sp1, DisplaySprite sp2)
-		{
-			return !(sp1 == sp2);
-		}
+            return sp1.CustomBit == sp2.CustomBit && sp1.ExtraBit == sp2.ExtraBit;
+        }
+        public static bool operator !=(DisplaySprite sp1, DisplaySprite sp2)
+        {
+            return !(sp1 == sp2);
+        }
 
-		#endregion
-        
-		public string GetTileLine()
-		{
-			if(UseText)
-				return Y.ToString("X") + X.ToString("X") + (2 + (ExtraBit ? 0x10 : 0) + (CustomBit ? 0x20 : 0)).ToString("X2") + " 0,0,*" + DisplayText + "*";
-			return Y.ToString("X") + X.ToString("X") + (2 + (ExtraBit ? 0x10 : 0) + (CustomBit ? 0x20 : 0)).ToString("X2") + " " + String.Join(" ", Tiles);
-		}
+        #endregion
 
-		public object Clone()
-		{
-			DisplaySprite s = new DisplaySprite();
-			s.CustomBit = this.CustomBit;
-			s.Description = this.Description;
-			s.ExtraBit = this.ExtraBit;
-			s.Tiles = new BindingList<Tile>();
+        public string GetTileLine()
+        {
+            if (UseText)
+                return Y_or_value.ToString("X") + X_or_index.ToString("X") + (2 + (ExtraBit ? 0x10 : 0) + (CustomBit ? 0x20 : 0)).ToString("X2") + " 0,0,*" + DisplayText + "*";
+            return Y_or_value.ToString("X") + X_or_index.ToString("X") + (2 + (ExtraBit ? 0x10 : 0) + (CustomBit ? 0x20 : 0)).ToString("X2") + " " + String.Join(" ", Tiles);
+        }
+
+        public object Clone()
+        {
+            DisplaySprite s = new DisplaySprite();
+            s.CustomBit = this.CustomBit;
+            s.Description = this.Description;
+            s.ExtraBit = this.ExtraBit;
+            s.Tiles = new BindingList<Tile>();
             foreach (Tile t in Tiles)
                 s.Tiles.Add((Tile)t.Clone());
-			s.X = this.X;
-			s.Y = this.Y;
+            s.X_or_index = this.X_or_index;
+            s.Y_or_value = this.Y_or_value;
             s.DisplayText = this.DisplayText;
             s.UseText = this.UseText;
-			return s;
-		}
+            return s;
+        }
 
         protected void SetPropertyValue<T>(ref T priv, T val, [CallerMemberName] string caller = "")
         {
-            bool pn = ReferenceEquals(priv, null);
-            bool vn = ReferenceEquals(val, null);
+            bool pn = priv == null;
+            bool vn = val == null;
 
             if (pn && vn)
                 return;
@@ -181,18 +264,18 @@ namespace CFG.Map16
     {
         public bool Equals(DisplaySprite x, DisplaySprite y)
         {
-            return x.X == y.X && x.Y == y.Y && x.ExtraBit == y.ExtraBit;
+            return x.X_or_index == y.X_or_index && x.Y_or_value == y.Y_or_value && x.ExtraBit == y.ExtraBit;
         }
 
         public int GetHashCode(DisplaySprite obj)
         {
-            return obj.X ^ obj.Y * (obj.ExtraBit ? -1 : 1);
+            return obj.X_or_index ^ obj.Y_or_value * (obj.ExtraBit ? -1 : 1);
         }
     }
 
     [DebuggerDisplay("{ToString()}")]
-	public class Tile : ICloneable
-	{
+    public class Tile : ICloneable
+    {
 
         [Newtonsoft.Json.JsonProperty("X offset")]
         public int XOffset { get; set; }
@@ -205,52 +288,52 @@ namespace CFG.Map16
         public bool Selected { get; set; }
 
         public Tile()
-		{ }
+        { }
 
-		public Tile(int x, int y, int num)
-		{
-			XOffset = x;
-			YOffset = y;
-			Map16Number = num;
-		}
-		
-		public override bool Equals(object obj)
-		{
-			if (Object.ReferenceEquals(this, obj))
-				return true;
-			if ((object)this == null)
-				return false;
+        public Tile(int x, int y, int num)
+        {
+            XOffset = x;
+            YOffset = y;
+            Map16Number = num;
+        }
 
-			Tile t = obj as Tile;
-			if ((object)t == null)
-				return false;
+        public override bool Equals(object obj)
+        {
+            if (Object.ReferenceEquals(this, obj))
+                return true;
+            if ((object)this == null)
+                return false;
 
-			return t.Map16Number == this.Map16Number && t.XOffset == this.XOffset && t.YOffset == this.YOffset;
-		}
+            Tile t = obj as Tile;
+            if ((object)t == null)
+                return false;
 
-		public override int GetHashCode()
-		{
-			return Map16Number + (XOffset << 13) + (YOffset << 20);
-		}
+            return t.Map16Number == this.Map16Number && t.XOffset == this.XOffset && t.YOffset == this.YOffset;
+        }
 
-		public override string ToString()
-		{
-			return XOffset + "," + YOffset + "," + Map16Number.ToString("X");
-		}
+        public override int GetHashCode()
+        {
+            return Map16Number + (XOffset << 13) + (YOffset << 20);
+        }
+
+        public override string ToString()
+        {
+            return XOffset + "," + YOffset + "," + Map16Number.ToString("X");
+        }
 
         const string regex = "(?<X>-?\\d+),(?<Y>-?\\d+),(?<M>[0-9a-fA-F]+)";
 
         public static Tile GetTile(string str)
-		{
-			var match = Regex.Match(str, regex);
+        {
+            var match = Regex.Match(str, regex);
             if (!match.Success)
                 return null;
 
-			int XOffset = Convert.ToInt32(match.Groups["X"].Value);
-			int YOffset = Convert.ToInt32(match.Groups["Y"].Value);
-			int Map16Number = Convert.ToInt32(match.Groups["M"].Value, 16);
+            int XOffset = Convert.ToInt32(match.Groups["X"].Value);
+            int YOffset = Convert.ToInt32(match.Groups["Y"].Value);
+            int Map16Number = Convert.ToInt32(match.Groups["M"].Value, 16);
             return new Tile(XOffset, YOffset, Map16Number);
-		}
+        }
 
         public static IEnumerable<Tile> GetTiles(string str)
         {
@@ -267,9 +350,9 @@ namespace CFG.Map16
             }
         }
 
-		public virtual object Clone()
-		{
-			return new Tile(XOffset, YOffset, Map16Number);
-		}
+        public virtual object Clone()
+        {
+            return new Tile(XOffset, YOffset, Map16Number);
+        }
     }
 }
