@@ -3,6 +3,8 @@
 
 #include "file_io.h"
 #include <cstring>
+#include <vector>
+#include <string>
 
 // use 16MB ROM size to avoid asar malloc/memcpy on 8MB of data per block.
 constexpr auto MAX_ROM_SIZE = 16 * 1024 * 1024;
@@ -41,28 +43,21 @@ struct tile {
     int x_offset = 0;
     int y_offset = 0;
     int tile_number = 0;
-    const char *text = nullptr;
-
-    ~tile();
+    std::string text{};
 };
 
 struct display {
-    char *description = nullptr;
-    size_t tile_count = 0;
-    tile *tiles = nullptr;
+    std::string description{};
+    std::vector<tile> tiles{};
     bool extra_bit = false;
     int x = 0;
     int y = 0;
-
-    ~display();
 };
 
 struct collection {
-    const char *name = nullptr;
+    std::string name{};
     bool extra_bit = false;
     unsigned char prop[12] = {0}; // why was this 4 again?
-
-    ~collection();
 };
 
 struct map8x8 {
@@ -116,14 +111,11 @@ struct sprite {
     const char *asm_file = nullptr;
     const char *cfg_file = nullptr;
 
-    size_t map_block_count = 0;
-    map16 *map_data = nullptr;
+    std::vector<map16> map_data{};
 
-    size_t display_count = 0;
-    display *displays = nullptr;
+    std::vector<display> displays{};
 
-    size_t collection_count = 0;
-    collection *collections = nullptr;
+    std::vector<collection> collections{};
 
     int sprite_type = 0; // 0 = Normal custom sprite, 1 = Extended custom sprite, 2 = Cluster custom sprite, 3 =
                          // Overworld custom sprite
